Remove conflicted symlink
A recent change to the upstream manifests added a symlink
that isn't actually useful with modern versions of the Cisco
plugin, and causes a duplicate define error if used in
conjuction with the OVS plugin.  This patch removes it.
We'll carry this patch temporarily due to time constraints,
but it will be dealt with upstream in a work-in-progress
refactoring of the Cisco plugin manifest.

Partial-Bug: #1276168