Fix missing plugin --config-file arguments
Ubuntu has broken init scripts for the ovs agent and
neutron-server in 14.04 in that they assume you're using the
ml2 plugin.  The server init script reads an /etc/default file
which can be modified to point to other plugin configs.  The
OVS agent has no such file (the config file is coded into the
init script).  A previous patch partly worked around the issue
by modifying the relevant files to prevent the ml2 config files
from being passed as --config-file arguments to the relevant
services.  However, that patch doesn't replace the ml2 config
files with the ovs config files, so the services start without
the config information contained in
/etc/neutron/plugins/openvswitch/ovs_neutron_plugin.ini.  This
patch corrects the problem by doing two things: first, it adds
the ovs_neutron_plugin.ini file to the ovs agent startup files.
Second, it points neutron-server to /etc/neutron/plugin.ini, which
is a symlink that points to whatever plugin file is actually being
used (thus solving the neutron-server problem not only for the
OVS plugin, but potentially other plugins as well).

Closes-Bug: #1322012